modified:   lib/data/services/openai_chat_service.dart 	modified:   lib/data/services/settings_service.dart 	modified:   lib/main.dart 	modified:   lib/presentation/providers/api_providers.dart 	modified:   lib/presentation/providers/chat_providers.dart 	modified:   lib/presentation/providers/settings_provider.dart 	modified:   lib/presentation/screens/settings_page.dart 	modified:   lib/presentation/widgets/right_settings_panel.dart

+++ lib/presentation/widgets/right_settings_panel.dart
@@ -119,8 +119,8 @@ class _RightSettingsPanelState extends ConsumerState<RightSettingsPanel> {
                       _buildIntInputSetting( // Top K
                          context, ref: ref,
                          label: 'Top K', // Label only
-                         value: settings.topK,
-                          onChanged: (val) => ref.read(settingsServiceProvider).setTopK(val),
+                         value: settings.topK.toInt(),
+                          onChanged: (val) => ref.read(settingsServiceProvider).setTopK(val.toDouble()),
                          minValue: 1,
                       ),
                        _buildIntInputSetting( // Max Output Tokens
